Add, update index implemented. Structure refactoring
