+++ UGSK.K3.Pulse/UGSK.K3.Pulse.csproj
@@ -190,6 +190,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Controllers.cs" />
+    <Compile Include="DapperDataStorage.cs" />
     <Compile Include="Data.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services.cs" />
