+++ UGSK.K3.Pulse/adm/product/product.module.js
@@ -4,6 +4,9 @@
     .config(['$resourceProvider', function ($resourceProvider) {
         // Don't strip trailing slashes from calculated URLs
         $resourceProvider.defaults.stripTrailingSlashes = false;
-    }]);
-    
+    }])
+    .constant('productModuleConfiguration',{
+        'IndexUrl':'/api/Index',
+        'CounterUrl': '/api/Counter'
+    });
 })();
