+++ UGSK.K3.Pulse/adm/product/products.html
@@ -23,7 +23,7 @@
                                     </td>
                                     <td><input type="number" class="form-control" min="0" max="1000" step="10" required="required" data-ng-model="item.Value"></td>
                                     <td>
-                                        <button type="button" class="btn btn-sm btn-info">Save</button>
+                                        <button type="button" class="btn btn-sm btn-info" ng-click="products.update(item)">Save</button>
                                     </tr>
                                 </tbody>
                             </table>
